Fixed issues with SDWebImage

+++ Example/Example/SDWebImage/SDWebImageDownloader.m
@@ -104,7 +104,7 @@ static NSString *const kCompletedCallbackKey = @"completed";
     return _downloadQueue.maxConcurrentOperationCount;
 }
 
-- (id <SDWebImageOperation>)downloadImageWithURL:(NSURL *)url options:(SDWebImageDownloaderOptions)options progress:(void (^)(NSInteger, NSInteger))progressBlock completed:(void (^)(UIImage *, NSData *, NSError *, BOOL))completedBlock {
+- (id <SDWebImageOperation>)downloadImageWithURL:(NSURL *)url options:(SDWebImageDownloaderOptions)options progress:(void (^)(NSUInteger, long long))progressBlock completed:(void (^)(UIImage *, NSData *, NSError *, BOOL))completedBlock {
     __block SDWebImageDownloaderOperation *operation;
     __weak SDWebImageDownloader *wself = self;
 
@@ -126,7 +126,7 @@ static NSString *const kCompletedCallbackKey = @"completed";
         }
         operation = [[SDWebImageDownloaderOperation alloc] initWithRequest:request
                                                                    options:options
-                                                                  progress:^(NSInteger receivedSize, NSInteger expectedSize) {
+                                                                  progress:^(NSUInteger receivedSize, long long expectedSize) {
                                                                       if (!wself) return;
                                                                       SDWebImageDownloader *sself = wself;
                                                                       NSArray *callbacksForURL = [sself callbacksForURL:url];
@@ -164,7 +164,7 @@ static NSString *const kCompletedCallbackKey = @"completed";
     return operation;
 }
 
-- (void)addProgressCallback:(void (^)(NSInteger, NSInteger))progressBlock andCompletedBlock:(void (^)(UIImage *, NSData *data, NSError *, BOOL))completedBlock forURL:(NSURL *)url createCallback:(void (^)())createCallback {
+- (void)addProgressCallback:(void (^)(NSUInteger, long long))progressBlock andCompletedBlock:(void (^)(UIImage *, NSData *data, NSError *, BOOL))completedBlock forURL:(NSURL *)url createCallback:(void (^)())createCallback {
     // The URL will be used as the key to the callbacks dictionary so it cannot be nil. If it is nil immediately call the completed block with no image or data.
     if (url == nil) {
         if (completedBlock != nil) {
